refactor: move the utils from component_tags to util.template_tag
