+++ docs/scripts/reference.py
@@ -49,7 +49,7 @@ from django.urls import URLPattern, URLResolver
 
 from django_components import ComponentVars, TagFormatterABC
 from django_components.component import Component
-from django_components.templatetags.component_tags import TagSpec
+from django_components.util.template_tag import TagSpec
 from django_components.util.misc import get_import_path
 
 # NOTE: This file is an entrypoint for the `gen-files` plugin in `mkdocs.yml`.
